Documentation and code contradiction fix tree.h
Documentation and code contradiction.
Maybe tree_get() should return T* value and not true value. And maybe it's T* and not T so it can return NULL if key is not found. Otherwise what value should it return if key is not found.
Same for tree_remove() it needs to return NULL if key not found.